Make a card component
This commit creates a new card component because it is used for multiple
times like in the hero twice and for the footer. So, it's best to make
it as its own component.

Also, I move the `_hero.scss` and the `_footer.scss` to the layout
folder because now the card is the actual component.