Refactor the "Show Password"
This commit fixes the issue where the "Show Password" checkbox is
stacked (text become two lines) on 320px width screen size. Now, the
"Show Password" checkbox uses eye icon and positioned it on the password
input.

This way, the "Password Length" can be aligned left and won't become two
lines which solve the problem.

I also create new class for the label styling so, that it's much easier
to distinguish the position styling between the icon label (eye icon) and the text label
(placeholder).