Change the alert messages
This commit makes the alert messages short and concise. This way, the
users can easily understand what the alert message is about. Also, add
the information about the valid password. It won't disappear (not in
alert message).

In addition to this changes, I make the alert messages exist on the page
until the users resubmit the form.